Remove duplicate implementation of the Osu playfield layer
